requirements.txt
appdirs==1.4.4
astroid==2.6.4
boto3==1.26.66
django-storages==1.13.2
bs4==0.0.1
chardet==5.1.0
charset-normalizer==2.0.11
configparser==5.0.2
distlib==0.3.1
Django==4.0
django-adminlte3==0.1.6
django-bootstrap4==21.2
django-ckeditor==6.2.0
django-sendgrid-v5==1.2.0
django-user-agents==0.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
gunicorn==20.1.0
google-auth==2.23.0
idna==3.3
importlib-metadata==2.1.3
jsonfield==2.1.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
mysqlclient==2.1.0
openpyxl==3.1.2
Pillow==9.3
pymysql==1.0.3
python-http-client==3.3.5
python-stdnum==1.17
requests==2.27.1
sendgrid==6.9.5
six==1.15.0
simplejson==3.18.1
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.8
user-agents==2.2.0
virtualenv==20.4.3
webdriver-manager==3.4.2
wrapt==1.12.1
zipp==3.7.0
django-environ==0.9.0
django-cors-headers==3.13.0
